+++ src/ClientData/CallstackData.cpp
@@ -93,45 +93,6 @@ std::vector<CallstackEvent> CallstackData::GetCallstackEventsOfTidInTimeRange(
   return callstack_events;
 }
 
-void CallstackData::ForEachCallstackEvent(
-    const std::function<void(const orbit_client_data::CallstackEvent&)>& action) const {
-  std::lock_guard<std::recursive_mutex> lock(mutex_);
-  for (const auto& [unused_tid, events] : callstack_events_by_tid_) {
-    for (const auto& [unused_timestamp, event] : events) {
-      action(event);
-    }
-  }
-}
-
-void CallstackData::ForEachCallstackEventInTimeRange(
-    uint64_t min_timestamp, uint64_t max_timestamp,
-    const std::function<void(const orbit_client_data::CallstackEvent&)>& action) const {
-  std::lock_guard<std::recursive_mutex> lock(mutex_);
-  ORBIT_CHECK(min_timestamp <= max_timestamp);
-  for (const auto& [unused_tid, events] : callstack_events_by_tid_) {
-    for (auto event_it = events.lower_bound(min_timestamp);
-         event_it != events.upper_bound(max_timestamp); ++event_it) {
-      action(event_it->second);
-    }
-  }
-}
-
-void CallstackData::ForEachCallstackEventOfTidInTimeRange(
-    uint32_t tid, uint64_t min_timestamp, uint64_t max_timestamp,
-    const std::function<void(const orbit_client_data::CallstackEvent&)>& action) const {
-  std::lock_guard<std::recursive_mutex> lock(mutex_);
-  ORBIT_CHECK(min_timestamp <= max_timestamp);
-  const auto& tid_and_events_it = callstack_events_by_tid_.find(tid);
-  if (tid_and_events_it == callstack_events_by_tid_.end()) {
-    return;
-  }
-  const auto& events = tid_and_events_it->second;
-  for (auto event_it = events.lower_bound(min_timestamp);
-       event_it != events.upper_bound(max_timestamp); ++event_it) {
-    action(event_it->second);
-  }
-}
-
 void CallstackData::AddCallstackFromKnownCallstackData(const CallstackEvent& event,
                                                        const CallstackData& known_callstack_data) {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
@@ -159,15 +120,6 @@ const CallstackInfo* CallstackData::GetCallstack(uint64_t callstack_id) const {
 bool CallstackData::HasCallstack(uint64_t callstack_id) const {
   std::lock_guard<std::recursive_mutex> lock(mutex_);
   return unique_callstacks_.contains(callstack_id);
-}
-
-void CallstackData::ForEachUniqueCallstack(
-    const std::function<void(uint64_t callstack_id, const CallstackInfo& callstack)>& action)
-    const {
-  std::lock_guard<std::recursive_mutex> lock(mutex_);
-  for (const auto& [callstack_id, callstack_ptr] : unique_callstacks_) {
-    action(callstack_id, *callstack_ptr);
-  }
 }
 
 std::shared_ptr<CallstackInfo> CallstackData::GetCallstackPtr(uint64_t callstack_id) const {
