+++ src/MizarData/CMakeLists.txt
@@ -9,7 +9,8 @@ add_library(MizarData STATIC)
 target_sources(MizarData PUBLIC 
          include/MizarData/BaselineAndComparison.h
          include/MizarData/MizarData.h
-         include/MizarData/MizarDataProvider.h)
+         include/MizarData/MizarDataProvider.h
+         include/MizarData/MizarPairedData.h)
 
 target_include_directories(MizarData PUBLIC ${CMAKE_CURRENT_LIST_DIR}/include) 
 
@@ -35,7 +36,7 @@ add_executable(MizarDataTests)
 target_sources(MizarDataTests PRIVATE 
                 BaselineAndComparisonTest.cpp
                 MizarDataTest.cpp
-                MizarDataWithSampledFunctionIdTest.cpp)
+                MizarPairedDataTest.cpp)
 
 target_link_libraries(MizarDataTests PRIVATE GrpcProtos
                                                 MizarData
